chore: duty/vacation 패키지 통합 (#21)
