--- src/main/java/com/example/server/duty/controller/DutyController.java
@@ -1,4 +1,0 @@
-package com.example.server.duty.controller;
-
-public class DutyController {
-}
