--- src/main/java/com/example/server/duty/repository/DutyRepository.java
@@ -1,4 +1,0 @@
-package com.example.server.duty.repository;
-
-public class DutyRepository {
-}
