--- src/main/java/com/example/server/duty/service/DutyService.java
@@ -1,4 +1,0 @@
-package com.example.server.duty.service;
-
-public class DutyService {
-}
