+++ src/main/java/com/example/server/schedule/Status.java
@@ -1,4 +1,4 @@
-package com.example.server.common;
+package com.example.server.schedule;
 
 public enum Status {
     PENDING, APPROVE, REJECT
