+++ src/main/java/com/example/server/schedule/duty/controller/DutyController.java
@@ -1,0 +1,4 @@
+package com.example.server.schedule.duty.controller;
+
+public class DutyController {
+}
