+++ src/main/java/com/example/server/schedule/duty/repository/DutyRepository.java
@@ -1,0 +1,4 @@
+package com.example.server.schedule.duty.repository;
+
+public class DutyRepository {
+}
