+++ src/main/java/com/example/server/schedule/duty/service/DutyService.java
@@ -1,0 +1,4 @@
+package com.example.server.schedule.duty.service;
+
+public class DutyService {
+}
