+++ src/main/java/com/example/server/schedule/vacation/controller/VacationController.java
@@ -1,0 +1,4 @@
+package com.example.server.schedule.vacation.controller;
+
+public class VacationController {
+}
