+++ src/main/java/com/example/server/schedule/vacation/enums/Reason.java
@@ -1,0 +1,5 @@
+package com.example.server.schedule.vacation.enums;
+
+public enum Reason {
+
+}
