+++ src/main/java/com/example/server/schedule/vacation/model/Vacation.java
@@ -1,25 +1,18 @@
-package com.example.server.duty.model;
+package com.example.server.schedule.vacation.model;
 
-
-import com.example.server.common.Status;
+import com.example.server.schedule.Status;
+import com.example.server.schedule.vacation.enums.Reason;
 import com.example.server.user.model.User;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
 
 import javax.persistence.*;
 import java.sql.Timestamp;
 
-@Entity
-@Table(name = "on_duty_tb")
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
 @Getter
-public class Duty {
+@Table(name = "vacation_tb")
+@Entity
+public class Vacation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -29,20 +22,24 @@ public class Duty {
     @JoinColumn(name = "user_id")
     private User user;
 
+    @Enumerated(EnumType.STRING)
     @Column(nullable = false)
-    private Timestamp dutyDate;
+    private Reason reason;
 
     @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
     private Status status;
 
-    @Column
+    @Column(nullable = false)
+    private Timestamp startDate;
+
+    @Column(nullable = false)
+    private Timestamp endDate;
+
     private Timestamp approvalDate;
 
     @CreationTimestamp
     private Timestamp createdDate;
-
-    @UpdateTimestamp
-    private Timestamp updatedDate;
 
     @PrePersist
     protected void onCreate() {
