+++ src/main/java/com/example/server/schedule/vacation/model/VacationInfo.java
@@ -1,4 +1,4 @@
-package com.example.server.vacation.model;
+package com.example.server.schedule.vacation.model;
 
 import com.example.server.user.model.User;
 import lombok.Getter;
