+++ src/main/java/com/example/server/schedule/vacation/repository/VacationRepository.java
@@ -1,0 +1,4 @@
+package com.example.server.schedule.vacation.repository;
+
+public class VacationRepository {
+}
