+++ src/main/java/com/example/server/schedule/vacation/service/VacationService.java
@@ -1,0 +1,4 @@
+package com.example.server.schedule.vacation.service;
+
+public class VacationService {
+}
