--- src/main/java/com/example/server/vacation/controller/VacationController.java
@@ -1,4 +1,0 @@
-package com.example.server.vacation.controller;
-
-public class VacationController {
-}
