--- src/main/java/com/example/server/vacation/enums/Reason.java
@@ -1,5 +1,0 @@
-package com.example.server.vacation.enums;
-
-public enum Reason {
-
-}
