--- src/main/java/com/example/server/vacation/repository/VacationRepository.java
@@ -1,4 +1,0 @@
-package com.example.server.vacation.repository;
-
-public class VacationRepository {
-}
