--- src/main/java/com/example/server/vacation/service/VacationService.java
@@ -1,4 +1,0 @@
-package com.example.server.vacation.service;
-
-public class VacationService {
-}
